Refactor configuration access: replace Plugin calls with fatal_error_sentinel function in BetterStack, Email, and Telegram services
